SIR Model for Math Project
Is able to make predictive models of pandemics and outbreaks according to the parameters you give it.

paramters to change for data is on SIR .py File line no. 45 and then line 49 can be used to change the timestep which basically is how many times it plots one single point for each equation